// src/assets/model/card/card.ts
export interface Card {
  Color?: string;
  CardType?: string;
  Subtype?: string;
  Name?: string;
  ManaCost?: number;
  PowerToughness?: string;
  Ability?: string;
  PlusMana?: string;
  PlusCardDraw?: string;
  Spirit?: string;
  Release?: string;
  Winnable?: string;
  CardNumber?: string;
  ImagePath?: string;
}